[REFORMATTING] [REFACTORING] General reformatting and refactoring

+++ schoolaccountingstudio-instruments/src/com/artiomastashonak/schoolaccountingstudio/invoice/Item.java
@@ -1,3 +1,3 @@
 package com.artiomastashonak.schoolaccountingstudio.invoice;
 
-public record Item(String measureUnit, int quantity, String code, String description, short vat, double price, double discount1, double discount2) { }
+public record Item(String measureUnit, int quantity, String code, String description, int vat, double price, double discount1, double discount2) { }
